rename the 'logger' package by 'journal' and move it into the 'internal' folder
